requirements.txt
Flask==2.3.3
pymongo==4.6.1
tensorflow-cpu==2.13.0
numpy==1.24.3
Pillow==10.0.1
geopy
